Implement filtered by categories

+++ src/main/java/com/shoptrap_ecommerce_backend/demo/dto/DtoFilter.java
@@ -2,12 +2,12 @@ package com.shoptrap_ecommerce_backend.demo.dto;
 
 import lombok.*;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class DtoFilter {
     private Long idCategory;
-    private BigInteger price;
+    private BigDecimal price;
     private Boolean discount;
 }
